rename application to app
