+++ app/main.go
@@ -1,4 +1,4 @@
-package application
+package app
 
 import (
 	"os"
